match-api/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
numpy==1.26.4
scikit-learn==1.5.2
sentence-transformers==3.0.1
pdfminer.six==20231228
python-docx==1.1.2
requests==2.32.3
tqdm==4.66.5
kaggle==1.6.17
tenacity==9.0.0
rapidfuzz==3.9.6
python-dotenv==1.0.1
joblib==1.4.2

